# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/base/code": {
            "get": {
                "description": "生成验证编码码和图片",
                "consumes": [
                    "application/json"
                ],
                "summary": "验证码接口",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "40001": {
                        "description": "验证码生成错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "登陆用户接口",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.UserLoginInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "40001": {
                        "description": "账户与密码不能为空",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "40003": {
                        "description": "参数绑定错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "注册用户接口",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.UserRegisterInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "40001": {
                        "description": "账户与密码不能为空",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "40002": {
                        "description": "参数绑定错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.UserLoginInfo": {
            "type": "object",
            "properties": {
                "captcha_digit": {
                    "type": "string"
                },
                "captcha_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "app.UserRegisterInfo": {
            "type": "object",
            "properties": {
                "captcha_digit": {
                    "type": "string"
                },
                "captcha_id": {
                    "type": "string"
                },
                "gender": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "integer"
                },
                "repassword": {
                    "type": "string"
                }
            }
        }
    }
}